chore: Add a way to build and install
Since `.gitignore` are a lot, its better to use a community updated list.
The list is downloaded and then copied over to the expected location (over 200 files!).

Fix when there is no .gitignore file (by ignoring the error!)

To build use `make build` and to install run `make install`.
This installs in the user's home dir hence no privileges are required.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>